sheet change detection, bound window title
A new class, SpriteSheetFileInformation was added to represent when a
sprite sheet is created, saved, or loaded.  The load/save commands were
modified to work with this for their success handlers.  The sprite sheet
VM now uses this instead of just the sprite sheet for its
NewSpriteSheetLoaded event.

The sprite sheet VM now has additional properties and internal event
handling to detect changes and allow for the associated file name to be
updated.  This also involved creating a SheetTitle property, which the
main window now binds to.