prompt to discard unsaved changes (close #9)
The sheet VM was modified to have an internal function to check and
prompt for unsaved changes, which is exposed through the
CheckUnsavedChangesHandler property.

This property is now used as the parameter for the new/load commands,
and is checked in the closing event for the main window (after changes
have been committed for any currently-focused element).