adding tooltips to fields (close #10)
There are static dictionaries, populated by the sprite sheet VM class
constructor, which use reflection to iterate over all public properties
on the sprite sheet/sheet types and add entries based on the Description
of the DisplayAttribute applied to those properties.

These dictionaries are exposed as ReadOnly properties and the ToolTip
attribute of each relevant text field is bound to the appropriate
dictionary entry.