smarter detection of when sprites can be moved
For the move commands, a pre-move delegate was added to the constructor
to raise NotifyPropertyChanging events for the VM's CanMoveUp and
CanMoveDown, and the post-move delegate was expanded to raise
NotifyPropertyChanged events.  The CanMoveUp and CanMoveDown properties
were changed to check the index of the current sprite in the collection.